+++ extensions/xcode/src/services/xcode-simulator.service.ts
@@ -199,10 +199,12 @@ export class XcodeSimulatorService {
       }
     }
     // Open URL in simulator
-    return execAsync(["xcrun", "simctl", "openurl", simulatorUDID ?? "booted", trimmedUrl].join(" ")).then(() => {
-      // Silently launch Simulator application
-      XcodeSimulatorService.launchSimulatorApplication();
-    });
+    return execAsync(["xcrun", "simctl", "openurl", simulatorUDID ?? "booted", `"${trimmedUrl}"`].join(" ")).then(
+      () => {
+        // Silently launch Simulator application
+        XcodeSimulatorService.launchSimulatorApplication();
+      }
+    );
   }
 
   /**